+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## v1.1.6 (2025-07-16)
 
 -   Added `"DeathTransfer"` to `VfxType`.
+-   Fixed a bug that the type of `TransformOffset` in editor should be `Transform` instead of `Trans`.
 
 ## v1.1.5 (2025-07-15)